lesson_10/Calculator/requirements.txt
pluggy==1.5.0
selenium==4.32.0
allure-pytest>=2.9.0
pytest>=7.0.0
urllib3[socks]==2.4.0